refactor(intelligence): split per-symbol regime state out of detector
detector.py (811 lines) bundled the stateful per-symbol accumulator
(_SymbolRegimeState: rolling price/indicator buffers + component-score
computation) with the MarketRegimeDetector classification orchestration. Move
_SymbolRegimeState into regime/symbol_state.py and re-export it from detector,
so the per-symbol state machine is separated from the classifier.

Imports apportioned cleanly (stateful_indicators + Online* indicators are
state-only; RegimeTransitionMatrix/RegimeState stay with the detector).
Re-export keeps the class identity stable for the one external importer
(test_detector_edges); 117 intelligence tests pass.
